refactor(sqs): creating SQS sublayer in messaging
