+++ tech-challenge-ecosystem/src/app.module.ts
@@ -1,11 +1,11 @@
 import { Module } from '@nestjs/common';
 import { TransactionController } from './api/controller/transaction.controller';
-import { TransactionProducerModule } from './messaging/producer/transaction/transaction.producer.module';
 import { ServicesModule } from './core/services/services.module';
 import { PayableController } from './api/controller/payable.controller';
+import { SQSModule } from './messaging/sqs/sqs.module';
 
 @Module({
-  imports: [ServicesModule, TransactionProducerModule],
+  imports: [ServicesModule, SQSModule],
   controllers: [TransactionController, PayableController],
 })
 export class AppModule {}
