+++ tech-challenge-ecosystem/src/messaging/sqs/sqs.module.ts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common';
+import { TransactionSQSModule } from './transaction/transaction.sqs.module';
+
+@Module({
+  imports: [TransactionSQSModule],
+  exports: [TransactionSQSModule],
+})
+export class SQSModule {}
